// react-app/src/misc/openapi.d.ts
export interface components {
    schemas: {
        AddNoteSourceRequest: {
            /** Format: int64 */
            noteId: number;
            /** Format: uri */
            link: string;
        };
// ...
        RemoveNoteSourceRequest: {
            /** Format: int64 */
            noteId: number;
            sourceId: string;
        };
        ReorderNoteSourcesRequest: {
            /** Format: int64 */
            noteId: number;
            sourceIds: string[];
        };
// ...
        SourcesResponse: {
            message?: string | null;
            sources: components["schemas"]["SourceItemResponse"][] | null;
        };
// ...
        UpdateNoteSourceRequest: {
            /** Format: int64 */
            noteId: number;
            sourceId: string;
            /** Format: uri */
            link: string;
            title?: string | null;
            description?: string | null;
            image?: string | null;
            showImage?: boolean;
        };
        UpdateNoteTagsRequest: {
            /** Format: int64 */
            id: number;
            tags: string[];
        };
        UpdateSectionRequest: {
            /** Format: int64 */
            id: number;
            displayName: string;
            color: string;
        };
        UserResponse: {
            id?: string | null;
            email?: string | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}